refactor: move loop runners to lib
