+++ indexer/lib/ipni-watcher.js
@@ -1,4 +1,5 @@
 import createDebug from 'debug'
+import timers from 'node:timers/promises'
 import { assertOkResponse } from './http-assertions.js'
 import { multiaddrToHttpUrl } from './vendored/multiaddr.js'
 
@@ -6,6 +7,35 @@ const debug = createDebug('spark-piece-indexer:observer')
 
 /** @import { ProviderToInfoMap, ProviderInfo } from './typings.js' */
 /** @import { RedisRepository as Repository } from './redis-repository.js' */
+
+/**
+ * @param {object} args
+ * @param {Repository} args.repository
+ * @param {number} args.minSyncIntervalInMs
+ * @param {AbortSignal} [args.signal]
+ */
+export async function runIpniSync ({ repository, minSyncIntervalInMs, signal }) {
+  while (!signal?.aborted) {
+    const started = Date.now()
+    try {
+      console.log('Syncing from IPNI')
+      const providers = await syncProvidersFromIPNI(repository)
+      console.log(
+        'Found %s providers, %s support(s) HTTP(s)',
+        providers.size,
+        Array.from(providers.values()).filter(p => p.providerAddress.match(/^https?:\/\//)).length
+      )
+    } catch (err) {
+      console.error('Cannot sync from IPNI.', err)
+      // TODO: log to Sentry
+    }
+    const delay = minSyncIntervalInMs - (Date.now() - started)
+    if (delay > 0) {
+      console.log('Waiting for %sms before the next sync from IPNI', delay)
+      await timers.setTimeout(delay)
+    }
+  }
+}
 
 /**
  * @returns {Promise<ProviderToInfoMap>}
